fixed wishlist crash && stop image sliding

+++ Abaya/Controller/ProductsVCs/WishListVC.swift
@@ -217,7 +217,7 @@ extension WishListVC:UICollectionViewDelegate,UICollectionViewDataSource,UIColle
         cell.lblName.text = dic .object(forKey: "brand") as?String
         cell.lbldetail.text = dic .object(forKey: "product_name") as?String
         
-        let x : Int = (dic .object(forKey: "product_current_price") as? NSInteger)!
+        let x : Double = (dic .object(forKey: "product_current_price") as? Double)!
         let stringValue = "\(x)"
         print(stringValue)
